+++ lib/tcc/tcc.hpp
@@ -1,9 +1,11 @@
 #pragma once
 
+#include <cstdint>
+
 #include "tcc/core/warnings.hpp"
 
 namespace tcc
 {
-using Integer  = int32_t;
-using UInteger = uint32_t;
+using Integer  = int64_t;
+using UInteger = uint64_t;
 }  // namespace tcc
